refactor: extract canvas recording system from activity.js
